- modify monetary values to decimal (better practise / precision ) - first wrking implementation of total method with discounts implemented and passing test

+++ checkout-kata/CheckoutLogic/ICheckout.cs
@@ -6,7 +6,7 @@ namespace checkout_kata.CheckoutLogic
     {
         List<Product> Basket { get; set; }
 
-        double GetTotalPrice();
+        decimal GetTotalPrice();
         void Scan(Product product);
     }
 }
